email-security-gateway/requirements.txt
fastapi
uvicorn
scikit-learn
joblib
python-dotenv
requests
tldextract
